config fixes
