+++ pkg/bundler/cmd.go
@@ -10,10 +10,8 @@
 package bundler
 
 import (
-	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
 	"strconv"
 
 	"github.com/golang-migrate/migrate/v4"
@@ -27,90 +25,12 @@ import (
 	"github.com/dhontecillas/hfw/pkg/obs/logs"
 )
 
-var (
-	ErrBundlerNoSection = errors.New("no 'bundler' section")
-	ErrBundlerParse     = errors.New("cannot parse 'bundler' section")
-	ErrBundlerValidate  = errors.New("cannot validate 'bundler' section")
-)
-
-type BundlerMigrationsConfig struct {
-	Collect bool     `json:"collect"`
-	Dst     string   `json:"dst"`
-	Scan    []string `json:"scan"`    // directories to scan for migrations
-	Migrate string   `json:"migrate"` // migrate version
-}
-
-func (c *BundlerMigrationsConfig) Validate() error {
-	// if Dst == "" there is no bundling enabled
-	if c.Dst != "" {
-		dst, err := filepath.Abs(c.Dst)
-		if err != nil {
-			return err
-		}
-		c.Dst = dst
-	}
-	// TODO: validate that the scan directoriese exist
-	return nil
-}
-
-type BundlerPackConfig struct {
-	// where to place the bundle files
-	Dst string `json:"dst"`
-	// from where the files are picked to be packed with the bundler:
-	Srcs []string `json:"srcs"`
-	// variant of the config file to be put inside the bundle:
-	Variant string `json:"variant"`
-}
-
-func (c *BundlerPackConfig) Validate() error {
-	if c.Dst != "" {
-		dst, err := filepath.Abs(c.Dst)
-		if err != nil {
-			return err
-		}
-		c.Dst = dst
-	}
-
-	return nil
-}
-
-type BundlerConfig struct {
-	Pack       BundlerPackConfig       `json:"pack"`
-	Migrations BundlerMigrationsConfig `json:"migrations"`
-}
-
-func (c *BundlerConfig) Validate() error {
-	if err := c.Pack.Validate(); err != nil {
-		return err
-	}
-	if err := c.Migrations.Validate(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func NewBundlerConfig(cldr config.ConfLoader) (*BundlerConfig, error) {
-	cldr, err := cldr.Section([]string{"bundler"})
-	if err != nil {
-		return nil, fmt.Errorf("%w: %s", ErrBundlerNoSection,
-			err.Error())
-	}
-	var cnf BundlerConfig
-	if err := cldr.Parse(&cnf); err != nil {
-		return nil, fmt.Errorf("%w: %s", ErrBundlerParse, err.Error())
-	}
-	if err := cnf.Validate(); err != nil {
-		return nil, fmt.Errorf("%w: %s", ErrBundlerValidate, err.Error())
-	}
-	return &cnf, nil
-}
-
 // ExecuteBundlerOperations parses the command line and environment
 // to find operations that the bundler should execute: collect migrations,
 // collect static files, run migrations, etc..
 // This function is a helper to be able to run those operations from
 // the same server executable file.
-func ExecuteBundlerOperations(conf *BundlerConfig, dbConf *db.Config, l logs.Logger) {
+func ExecuteBundlerOperations(conf *config.BundlerConfig, dbConf *db.Config, l logs.Logger) {
 
 	if conf.Migrations.Collect {
 		if err := UpdateMigrations(conf.Migrations.Dst, conf.Migrations.Scan, l); err != nil {
@@ -147,7 +67,7 @@ func ExecuteBundlerOperations(conf *BundlerConfig, dbConf *db.Config, l logs.Log
 
 // ApplyMigrationsFromConfig reads the migrations firectory from config and
 // applies them.
-func ApplyMigrationsFromConfig(conf *BundlerMigrationsConfig, dbConf *db.Config, l logs.Logger) error {
+func ApplyMigrationsFromConfig(conf *config.BundlerMigrationsConfig, dbConf *db.Config, l logs.Logger) error {
 	if conf.Migrate == "" {
 		l.Info("no migration to apply", nil)
 		return nil
